@@ -57,6 +57,7 @@
  */
 
 import KripkeStructure
+import Dispatch
 
 public final class InMemoryKripkeStructure: MutableKripkeStructure {
 
@@ -71,17 +72,23 @@ public final class InMemoryKripkeStructure: MutableKripkeStructure {
     var allStates: [Int64: (KripkeStatePropertyList, Bool, Set<KripkeEdge>)] = [:]
 
     public var acceptingStates: AnySequence<KripkeState> {
-        AnySequence(states.filter { $0.edges.isEmpty })
+        DispatchQueue.global(qos: .userInteractive).sync {
+            AnySequence(states.filter { $0.edges.isEmpty })
+        }
     }
 
     public var initialStates: AnySequence<KripkeState> {
-        AnySequence(states.filter { $0.isInitial })
+        DispatchQueue.global(qos: .userInteractive).sync {
+            AnySequence(states.filter { $0.isInitial })
+        }
     }
 
     public var states: AnySequence<KripkeState> {
-        AnySequence(allStates.keys.map {
-            try! self.state(for: $0)
-        })
+        DispatchQueue.global(qos: .userInteractive).sync {
+            AnySequence(allStates.keys.map {
+                try! self.state(for: $0)
+            })
+        }
     }
 
     init(identifier: String) {
@@ -91,31 +98,40 @@ public final class InMemoryKripkeStructure: MutableKripkeStructure {
     init(identifier: String, states: Set<KripkeState>) throws {
         self.identifier = identifier
         for state in states {
-            let id = try self.add(state.properties, isInitial: state.isInitial)
+            let (id, _) = try self.add(state.properties, isInitial: state.isInitial)
             for edge in state.edges {
                 try self.add(edge: edge, to: id)
             }
         }
     }
 
-    public func add(_ propertyList: KripkeStatePropertyList, isInitial: Bool) throws -> Int64 {
-        let id = try id(for: propertyList)
-        if nil == allStates[id] {
-            allStates[id] = (propertyList, isInitial, [])
+    public func add(_ propertyList: KripkeStatePropertyList, isInitial: Bool) throws -> (Int64, Bool) {
+        try DispatchQueue.global(qos: .userInteractive).sync {
+            let id = try id(for: propertyList)
+            let inCycle = nil != allStates[id]
+            if !inCycle {
+                allStates[id] = (propertyList, isInitial, [])
+            }
+            return (id, inCycle)
         }
-        return id
     }
 
     public func add(edge: KripkeEdge, to id: Int64) throws {
-        allStates[id]?.2.insert(edge)
+        _ = DispatchQueue.global(qos: .userInteractive).sync {
+            allStates[id]?.2.insert(edge)
+        }
     }
 
     public func markAsInitial(id: Int64) throws {
-        self.allStates[id]?.1 = true
+        DispatchQueue.global(qos: .userInteractive).sync {
+            self.allStates[id]?.1 = true
+        }
     }
 
     public func exists(_ propertyList: KripkeStatePropertyList) throws -> Bool {
-        return nil != ids[propertyList]
+        DispatchQueue.global(qos: .userInteractive).sync {
+            return nil != ids[propertyList]
+        }
     }
 
     public func data(for propertyList: KripkeStatePropertyList) throws -> (Int64, KripkeState) {
@@ -124,24 +140,28 @@ public final class InMemoryKripkeStructure: MutableKripkeStructure {
     }
 
     public func id(for propertyList: KripkeStatePropertyList) throws -> Int64 {
-        if let id = ids[propertyList] {
+        DispatchQueue.global(qos: .userInteractive).sync {
+            if let id = ids[propertyList] {
+                return id
+            }
+            let id = latestId
+            latestId += 1
+            ids[propertyList] = id
             return id
         }
-        let id = latestId
-        latestId += 1
-        ids[propertyList] = id
-        return id
     }
 
     public func state(for id: Int64) throws -> KripkeState {
-        guard let (plist, isInitial, edges) = allStates[id] else {
-            fatalError("State does not exist")
+        DispatchQueue.global(qos: .userInteractive).sync {
+            guard let (plist, isInitial, edges) = allStates[id] else {
+                fatalError("State does not exist")
+            }
+            let state = KripkeState(isInitial: isInitial, properties: plist)
+            for edge in edges {
+                state.addEdge(edge)
+            }
+            return state
         }
-        let state = KripkeState(isInitial: isInitial, properties: plist)
-        for edge in edges {
-            state.addEdge(edge)
-        }
-        return state
     }
 
 
